Extract useful waterfall methods to utils.
