+++ lib/runner/extensions/waterfall.command.js
@@ -1,65 +1,7 @@
 var _ = require('lodash'),
     Cursor = require('../cursor'),
-    VariableScope = require('postman-collection').VariableScope,
-    { prepareVaultVariableScope } = require('../util'),
-
-    prepareLookupHash,
-    extractSNR,
-    getIterationData;
-
-/**
- * Returns a hash of IDs and Names of items in an array
- *
- * @param {Array} items -
- * @returns {Object}
- */
-prepareLookupHash = function (items) {
-    var hash = {
-        ids: {},
-        names: {},
-        obj: {}
-    };
-
-    _.forEach(items, function (item, index) {
-        if (item) {
-            item.id && (hash.ids[item.id] = index);
-            item.name && (hash.names[item.name] = index);
-        }
-    });
-
-    return hash;
-};
-
-extractSNR = function (executions, previous) {
-    var snr = previous || {};
-
-    _.isArray(executions) && executions.forEach(function (execution) {
-        _.has(_.get(execution, 'result.return'), 'nextRequest') && (
-            (snr.defined = true),
-            (snr.value = execution.result.return.nextRequest)
-        );
-    });
-
-    return snr;
-};
-
-/**
- * Returns the data for the given iteration
- *
- * @function getIterationData
- * @param {Array} data - The data array containing all iterations' data
- * @param {Number} iteration - The iteration to get data for
- * @return {Any} - The data for the iteration
- */
-getIterationData = function (data, iteration) {
-    // if iteration has a corresponding data element use that
-    if (iteration < data.length) {
-        return data[iteration];
-    }
-
-    // otherwise use the last data element
-    return data[data.length - 1];
-};
+    { prepareVaultVariableScope, prepareVariablesScope, prepareLookupHash,
+        extractSNR, getIterationData } = require('../util');
 
 /**
  * Adds options
@@ -71,19 +13,8 @@ module.exports = {
     init: function (done) {
         var state = this.state;
 
-        // ensure that the environment, globals and collectionVariables are in VariableScope instance format
-        state.environment = VariableScope.isVariableScope(state.environment) ? state.environment :
-            new VariableScope(state.environment);
-        state.globals = VariableScope.isVariableScope(state.globals) ? state.globals :
-            new VariableScope(state.globals);
-        state.vaultSecrets = VariableScope.isVariableScope(state.vaultSecrets) ? state.vaultSecrets :
-            new VariableScope(state.vaultSecrets);
-        state.collectionVariables = VariableScope.isVariableScope(state.collectionVariables) ?
-            state.collectionVariables : new VariableScope(state.collectionVariables);
-        state._variables = VariableScope.isVariableScope(state.localVariables) ?
-            state.localVariables : new VariableScope(state.localVariables);
-
-        // prepare the vault variable scope
+        // prepare the vault variable scope and other variables
+        prepareVariablesScope(state);
         prepareVaultVariableScope(state.vaultSecrets);
 
         // ensure that the items and iteration data set is in place
