+++ lib/runner/util.js
@@ -1,4 +1,5 @@
 var { Url, UrlMatchPatternList, VariableList } = require('postman-collection'),
+    VariableScope = require('postman-collection').VariableScope,
     sdk = require('postman-collection'),
     _ = require('lodash'),
 
@@ -236,6 +237,86 @@ module.exports = {
 
         // mark the scope as a vault variable scope
         scope.__vaultVariableScope = true;
+    },
+
+    /**
+     * ensure that the environment, globals and collectionVariables are in VariableScope instance format
+     * @param {*} state application state object.
+     */
+    prepareVariablesScope (state) {
+        state.environment = VariableScope.isVariableScope(state.environment) ? state.environment :
+            new VariableScope(state.environment);
+        state.globals = VariableScope.isVariableScope(state.globals) ? state.globals :
+            new VariableScope(state.globals);
+        state.vaultSecrets = VariableScope.isVariableScope(state.vaultSecrets) ? state.vaultSecrets :
+            new VariableScope(state.vaultSecrets);
+        state.collectionVariables = VariableScope.isVariableScope(state.collectionVariables) ?
+            state.collectionVariables : new VariableScope(state.collectionVariables);
+        state._variables = VariableScope.isVariableScope(state.localVariables) ?
+            state.localVariables : new VariableScope(state.localVariables);
+    },
+
+    /**
+ * Returns a hash of IDs and Names of items in an array
+ *
+ * @param {Array} items -
+ * @returns {Object}
+ */
+    prepareLookupHash (items) {
+        var hash = {
+            ids: {},
+            names: {},
+            obj: {}
+        };
+
+        _.forEach(items, function (item, index) {
+            if (item) {
+                item.id && (hash.ids[item.id] = index);
+                item.name && (hash.names[item.name] = index);
+            }
+        });
+
+        return hash;
+    },
+
+
+    /**
+     * Extract set next request from the execution.
+     *
+     * @function getIterationData
+     * @param {Array} executions - The prerequests or the tests of an item's execution.
+     * @param {Object} previous - If extracting the tests request then prerequest's snr.
+     * @return {Any} - The Set Next Request
+     */
+    extractSNR (executions, previous) {
+        var snr = previous || {};
+
+        _.isArray(executions) && executions.forEach(function (execution) {
+            _.has(_.get(execution, 'result.return'), 'nextRequest') && (
+                (snr.defined = true),
+                (snr.value = execution.result.return.nextRequest)
+            );
+        });
+
+        return snr;
+    },
+
+    /**
+     * Returns the data for the given iteration
+     *
+     * @function getIterationData
+     * @param {Array} data - The data array containing all iterations' data
+     * @param {Number} iteration - The iteration to get data for
+     * @return {Any} - The data for the iteration
+     */
+    getIterationData (data, iteration) {
+    // if iteration has a corresponding data element use that
+        if (iteration < data.length) {
+            return data[iteration];
+        }
+
+        // otherwise use the last data element
+        return data[data.length - 1];
     },
 
     /**
